Update patch set 3
Patch Set 3:

> While we are on this let report a nasty bug in Profiles which
 > wasn't present in CM13. Maybe is related to this commit hre.
 > To reproduce: Set Lock screen disabled in the active profile and
 > reboot. You end up with black screen, no launcher, only SystemUI is
 > running, you can only expand status bar and that's pretty much all
 > you can do. Seems like security issue blocking launcher even
 > built-in Trebuchet. The only way to get out of this is to delete
 > /data/system/profiles.xml and reboot. And you loose a bunch of your
 > settings on top of everything.
 > BTW this profiles feature is the only way I know of using
 > fingerprint without lock screen on your way.

I can confirm that this is a bug that exists.

In an effort to spare anyone else the pain of re/setting/s. Depending on how you have your profile triggers set up, it may be possible to trigger a different profile in which the lock screen is not disabled. In my case I was able to get back in by disabling wifi - though I did have to let it sit for a minute or two before triggering the profile change worked.

 > While we are on this let report a nasty bug in Profiles which
 > wasn't present in CM13. Maybe is related to this commit hre.
 > To reproduce: Set Lock screen disabled in the active profile and
 > reboot. You end up with black screen, no launcher, only SystemUI is
 > running, you can only expand status bar and that's pretty much all
 > you can do. Seems like security issue blocking launcher even
 > built-in Trebuchet. The only way to get out of this is to delete
 > /data/system/profiles.xml and reboot. And you loose a bunch of your
 > settings on top of everything.
 > BTW this profiles feature is the only way I know of using
 > fingerprint without lock screen on your way.

I can confirm that this is a bug that exists.

In an effort to spare anyone else the pain of re/setting/s. Depending on how you have your profile triggers set up, it may be possible to trigger a different profile in which the lock screen is not disabled. In my case, I was able to get back in by disabling my wifi. I did have to let it sit for a minute or two before triggering the profile change worked.

Patch-set: 3
Reviewer: Joey Marsh <17023@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0